Updated .NET SDK version and refactored Makefile
The .NET SDK version has been updated in the global.json file. In the Makefile, all instances of 'dotnet' have been replaced with a variable for better flexibility. The DOTNET_RID shell command has also been removed. A newline was added at the end of dotnet-tools.json to adhere to best practices.